requirements.txt
openai==0.27.8
langchain==0.0.228
python-dotenv==1.0.0
flask==2.3.3
requests==2.31.0
